[*][GameEngine] Sprite windows, onclick zone

+++ src/gameEngine/Position2D.cpp
@@ -65,7 +65,7 @@ int Position2D::getY() const
   return (this->_vector2.y);
 }
 
-sf::Vector2<int> Position2D::getVector() const
+sf::Vector2<float> Position2D::getVector() const
 {
   return (this->_vector2);
 }
